refactor(bot): move adapter interfaces to adapter
