+++ booking-telegram-bot/adapters/tgsnake-adapter/src/index.ts
@@ -1,2 +1,3 @@
 export * from './module/index.js'
 export * from './services/index.js'
+export * from './interfaces/index.js'
