+++ booking-telegram-bot/adapters/tgsnake-adapter/src/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './tgsnake.interfaces.js'
