+++ booking-telegram-bot/adapters/tgsnake-adapter/src/services/index.ts
@@ -1,2 +1,1 @@
-export * from './tgsnake-adapter.interfaces.js'
 export * from './tgsnake-adapter.service.js'
