--- booking-telegram-bot/contexts/telegram-bot/infrastructure-module/src/processors/bot-listen.interfaces.ts
@@ -1,13 +1,0 @@
-type TelegramBotFormattedContextType = {
-  userId: bigint
-  messageText: string
-  accessHash: bigint
-  messageId: number
-  chatId: bigint
-  replyMessage: (text: string) => Promise<void>
-}
-
-type TelegramBotFormattedContextKeyType = keyof TelegramBotFormattedContextType
-
-export type { TelegramBotFormattedContextType }
-export type { TelegramBotFormattedContextKeyType }
