+++ booking-telegram-bot/contexts/telegram-bot/infrastructure-module/src/processors/index.ts
@@ -1,2 +1,1 @@
 export * from './bot-listen.processor.js'
-export * from './bot-listen.interfaces.js'
